+++ src/ui/composites/cells/AssigneesCell.ts
@@ -1,10 +1,10 @@
-import { AvatarStack } from '../../primitives/AvatarStack'
+import { AvatarStack, type AvatarPerson } from '../../primitives/AvatarStack'
 
 export class AssigneesCell {
   el: HTMLTableCellElement
 
-  constructor(parentRow: HTMLElement, assignees: string[]) {
+  constructor(parentRow: HTMLElement, people: AvatarPerson[]) {
     this.el = parentRow.createEl('td', { cls: 'pm-table-cell pm-table-cell-assignees' })
-    new AvatarStack(this.el).setNames(assignees).setMax(3)
+    new AvatarStack(this.el).setPeople(people).setMax(3)
   }
 }